backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic[email]==2.10.4
pydantic-settings==2.7.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
python-docx==1.1.2
chromadb==0.6.3
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
tiktoken==0.8.0
httpx==0.28.1
